requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
scipy>=1.10.0
pandas>=2.0.0
cupy-cuda12x>=12.0.0
numba>=0.57.0
